# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online hautarzt API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/enquires/download-conclusion": {
            "post": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Download enquire's conclusion",
                "description": "Download enquire's conclusion",
                "responses": {
                    "200": {
                        "description": "Verification code has been verified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "conclusion": {
                                            "type": "string",
                                            "example": "Ri0xLjMKMSAwIG9iago8PCAvVHlwZSAvQ2F0YWxvZwovT3V0bGluZXMgMiAwIFIKL1BhZ2VzIDMgMC"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Something went wrong, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/enquires/{id}/send-sms": {
            "get": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Send SMS verification code",
                "description": "Send SMS verification code",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "An enquire's identificator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification code has been sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "SMS token was sent."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "There are some validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Validation error",
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "authy": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Failed to send verification code."
                                                    }
                                                }
                                            },
                                            "type": "object",
                                            "format": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "No query results for model [App\\Models\\Enquire]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Something went wrong, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/enquires/{id}/update-conclusion": {
            "patch": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Update an enquire's conclusion by id",
                "description": "Update an enquire's conclusion by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "An enquire's identificator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "conclusion"
                                ],
                                "properties": {
                                    "conclusion": {
                                        "description": "An enquire's conclusion",
                                        "format": "string",
                                        "example": "test conclusion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conclusion has been succesfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EnquireResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Current user has not permissions to do this action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enquiere not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "No query results for model [App\\Models\\Enquire]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Something went wrong, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarbonResource": {
                "title": "Carbon resource",
                "description": "Resource for datetime representation",
                "properties": {
                    "date": {
                        "format": "string",
                        "example": "2019-12-11 12:25:00.000000"
                    },
                    "timezone_type": {
                        "format": "int64",
                        "example": 3
                    },
                    "timezone": {
                        "format": "string",
                        "example": "Europe/Berlin"
                    }
                },
                "type": "object"
            },
...
            "EnquireResource": {
                "title": "Enquire resource",
                "description": "Resource for an enquire representation",
                "properties": {
                    "id": {
                        "schema": "Enquire",
                        "title": "ID",
                        "description": "An enquire's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "first_name": {
                        "schema": "Enquire",
                        "title": "First name",
                        "description": "An enquire customer's first name",
                        "format": "string",
                        "example": "Davide"
                    },
                    "last_name": {
                        "schema": "Enquire",
                        "title": "Last name",
                        "description": "An enquire customers's last name",
                        "format": "string",
                        "example": "Donghi"
                    },
                    "gender": {
                        "schema": "Enquire",
                        "title": "Gender",
                        "description": "An enquire customer's gender",
                        "format": "string",
                        "example": "MALE"
                    },
                    "date_of_birth": {
                        "schema": "Enquire",
                        "title": "Date of birth",
                        "description": "Date of birth representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "created_at": {
                        "schema": "Enquire",
                        "title": "Created at",
                        "description": "Created datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "updated_at": {
                        "schema": "Enquire",
                        "title": "Updated at",
                        "description": "Last updated datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "phone_number": {
                        "schema": "Enquire",
                        "title": "Phone number",
                        "description": "An enquire customer's phone number",
                        "format": "string",
                        "example": "+38 024-548-58-55"
                    },
                    "email": {
                        "schema": "Enquire",
                        "title": "E-mail",
                        "description": "An enquire customer's e-mail",
                        "format": "string",
                        "example": "test@gmail.com"
                    },
                    "conclusion": {
                        "schema": "Enquire",
                        "title": "Conclusion",
                        "description": "A doctors's conclusion",
                        "format": "string",
                        "example": "Everythink will be fine"
                    },
                    "conclusion_created_at": {
                        "schema": "Enquire",
                        "title": "Conclusion created at",
                        "description": "Conclusion created datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "status": {
                        "schema": "Enquire",
                        "title": "Status",
                        "description": "An enquire's status",
                        "format": "string",
                        "example": "UNREAD"
                    },
                    "is_paid": {
                        "schema": "Enquire",
                        "title": "Is paid",
                        "description": "Is an enquire paid",
                        "format": "boolean",
                        "example": "true"
                    },
                    "location": {
                        "schema": "Enquire",
                        "$ref": "#/components/schemas/LocationResource"
                    },
                    "answers": {
                        "schema": "Enquire",
                        "title": "Answers",
                        "description": "An enquire's answers",
                        "items": {
                            "type": "object",
                            "$ref": "#/components/schemas/EnquireAnswerResource"
                        }
                    }
                },
                "type": "object"
            },
            "EnquireAnswerResource": {
                "title": "Location resource",
                "description": "Resource for a location representation",
                "properties": {
                    "id": {
                        "schema": "EnquireAnswer",
                        "title": "ID",
                        "description": "An enquire answers's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "message": {
                        "schema": "EnquireAnswer",
                        "description": "An enquire answers's message",
                        "$ref": "#/components/schemas/MessageResource"
                    },
                    "message_option": {
                        "schema": "EnquireAnswer",
                        "description": "An enquire answers's selected message option",
                        "$ref": "#/components/schemas/MessageOptionResource"
                    },
                    "value": {
                        "schema": "EnquireAnswer",
                        "title": "Value",
                        "description": "An enquire answers's value",
                        "format": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
            "LocationResource": {
                "title": "Location resource",
                "description": "Resource for a location representation",
                "properties": {
                    "id": {
                        "schema": "Location",
                        "title": "ID",
                        "description": "A locations's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "country": {
                        "schema": "Location",
                        "title": "A location's country",
                        "description": "A location's country",
                        "format": "string",
                        "example": "USA"
                    },
                    "city": {
                        "schema": "Location",
                        "title": "A location's city",
                        "description": "A location's city",
                        "format": "string",
                        "example": "New York"
                    },
                    "state": {
                        "schema": "Location",
                        "title": "A location's state",
                        "description": "A location's state",
                        "format": "string",
                        "example": "New York"
                    },
                    "postal_code": {
                        "schema": "Location",
                        "title": "A location's postal code",
                        "description": "A location's postal code",
                        "format": "string",
                        "example": "12345"
                    },
                    "address": {
                        "schema": "Location",
                        "title": "A location's address",
                        "description": "A location's address",
                        "format": "string",
                        "example": "address"
                    },
                    "latitude": {
                        "schema": "Location",
                        "title": "A location's latitude",
                        "description": "A location's latitude",
                        "format": "double",
                        "example": 5.123,
                        "nullable": true
                    },
                    "longitude": {
                        "schema": "Location",
                        "title": "A location's longitude",
                        "description": "A location's longitude",
                        "format": "double",
                        "example": 8.123,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MessageResource": {
                "title": "Message resource",
                "description": "Resource for a message representation",
                "properties": {
                    "id": {
                        "schema": "Message",
                        "title": "ID",
                        "description": "A message's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "title": {
                        "schema": "Message",
                        "title": "A message's title",
                        "description": "A message's title",
                        "format": "string",
                        "example": "How are you?"
                    },
                    "content": {
                        "schema": "Message",
                        "title": "A message's content",
                        "description": "A message's content",
                        "format": "string",
                        "example": "How are you?"
                    },
                    "questioner": {
                        "schema": "Message",
                        "title": "A message's questioner",
                        "description": "A message's questioner",
                        "format": "string",
                        "example": "Doctor"
                    },
                    "type": {
                        "schema": "Message",
                        "title": "A message's type",
                        "description": "A message's type",
                        "format": "string",
                        "example": "simple"
                    },
                    "button": {
                        "schema": "Message",
                        "title": "A message's button",
                        "description": "A message's button",
                        "format": "string",
                        "example": "Click here!"
                    },
                    "next_message_id": {
                        "schema": "Message",
                        "title": "ID",
                        "description": "Next message identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "options": {
                        "schema": "Message",
                        "title": "Options",
                        "description": "Message's options",
                        "items": {
                            "type": "object",
                            "$ref": "#/components/schemas/MessageOptionResource"
                        }
                    }
                },
                "type": "object"
            },
            "MessageOptionResource": {
                "title": "Message option resource",
                "description": "Resource for a message option representation",
                "properties": {
                    "id": {
                        "schema": "MessageOption",
                        "title": "ID",
                        "description": "A message option's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "value": {
                        "schema": "MessageOption",
                        "title": "A message option's value",
                        "description": "A message option's value",
                        "format": "string",
                        "example": "value"
                    },
                    "message": {
                        "schema": "MessageOption",
                        "$ref": "#/components/schemas/MessageResource"
                    },
                    "next_message_id": {
                        "schema": "MessageOption",
                        "title": "ID",
                        "description": "Next message identificator",
                        "format": "int64",
                        "example": "5"
                    }
                },
                "type": "object"
            },
...
        }
    }
}